requirements.txt
langchain-core>=0.1.0
langchain-google-genai>=0.0.5
langgraph>=0.0.15
python-dotenv>=1.0.0
google-generativeai>=0.3.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
yfinance>=0.2.36
beautifulsoup4>=4.12.0
pytest>=7.4.0
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.0
